Refactoring some of the view method into helper
